edge-agent/requirements.txt
kafka-python
minio
pymilvus
opencv-python-headless
ultralytics
torch
transformers
